+++ crates/runtime/src/runtime_modules.rs
@@ -13,6 +13,7 @@
 pub(crate) fn source(specifier: &str) -> Option<&'static str> {
     match specifier {
         "runtime:process" => Some(include_str!("runtime_modules/process.js")),
+        "runtime:path" => Some(include_str!("runtime_modules/path.js")),
         _ => None,
     }
 }